refactor: move useGetBannerQuery outside of package
