+++ client/src/data-provider/Misc/index.ts
@@ -1,0 +1,1 @@
+export * from './queries';
