+++ client/src/data-provider/index.ts
@@ -2,6 +2,7 @@ export * from './Auth';
 export * from './Agents';
 export * from './Files';
 export * from './Messages';
+export * from './Misc';
 export * from './Tools';
 export * from './connection';
 export * from './mutations';
