+++ packages/data-provider/src/react-query/react-query-service.ts
@@ -453,14 +453,3 @@ export const useGetCustomConfigSpeechQuery = (
     },
   );
 };
-
-export const useGetBannerQuery = (
-  config?: UseQueryOptions<t.TBannerResponse>,
-): QueryObserverResult<t.TBannerResponse> => {
-  return useQuery<t.TBannerResponse>([QueryKeys.banner], () => dataService.getBanner(), {
-    refetchOnWindowFocus: false,
-    refetchOnReconnect: false,
-    refetchOnMount: false,
-    ...config,
-  });
-};
